smart_kitchen_vn/requirements.txt
Flask>=2.0
flask-socketio>=5.0
eventlet
paho-mqtt
twilio
python-dotenv
